Complete product
The application begins with a standard name entry screen. The user must enter something, proceeding without anything entered will result in an exception being thrown. The application then transitions to a second screen where all food objects are dynamically generated. This allows the owner to move and add items to the list as they please. I did much more than what was required for the project.